sandbox: hostfile: add feature controller support
Keeping the tradition of making the sandbox more complex than it needs
to in order to exercise more parts of barebox, let's allow hostfiles to
be feature controllers: This allows specifying optional hostfiles in the
DT: If the hostfile is unavailable, the nodes pointing at the hostfile
can be gated by it, so they behave as if they were disabled.

This is useful for the stickypage, which results in a number of ugly
errors whenever it's unavailable.

Signed-off-by: Ahmad Fatoum <ahmad@a3f.at>
Link: https://lore.barebox.org/20230424121805.150434-7-ahmad@a3f.at
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>