ARM: vexpress: Disable non working devices
Disable some non working devices in the vexpress dts files to
avoid error messages during runtime. vexpress_defconfig has the
drivers for these devices disabled, but with multi_v7_defconfig
the driver are enabled. To make vexpress more usable with
multi_v7_defconfig disable the non working devices in the device
tree.

/timer@1e000600 doesn't work because our arm,cortex-a9-twd-timer
implementation needs a clock which the dts doesn't provide. We don't
need this as we use another timer anyway.

/bus@40000000/motherboard-bus@40000000/leds/ and /leds/ are a gpio-leds
devices which need a GPIO driver which doesn't exist in barebox

/bus@40000000/motherboard-bus@40000000/iofpga@7,00000000/compact-flash@1a000/
/bus@8000000/motherboard-bus@8000000/iofpga-bus@300000000/compact-flash@1a0000/
are ata-generic devices. The driver times out probing the device. While
it doesn't issue an error message it delays starting up the board for
several seconds.

Link: https://lore.barebox.org/20230425133535.273920-1-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>