defconfigs: usb-a9g20: update to a working version
The defconfig needs to be updated to get a usable barebox:

We need bigger malloc because of the increasing kernel size, soft ECC so
NAND will work, ifup and relocatable for easier development. Fix the
prompt while here. Sadly, because of size restrictions, FAT, NFS, menus,
netconsole, and SHA1 have to go.

Signed-off-by: Wolfram Sang <wsa@kernel.org>
Link: https://lore.barebox.org/20230427184718.105393-3-wsa@kernel.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>