MAKEALL: Do not print sizes
The size of the barebox proper binary is not very useful to print,
the binary will be compressed which decreases its size and the PBL
will be added to it which then adds to the resulting size again.

Printing the sizes will only show what we all know already: Software
gets bigger over time when new features are added and more corner
cases are handled. Just drop printing of the sizes.

Link: https://lore.barebox.org/20230502073925.266116-5-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>