ARM: rockchip: bbu: Fix getting active boot slot
With some TF-A versions the IRAM where the boot information is stored
cannot be accessed in normal mode. We therefore copy the information
to another place before starting the TF-A. This copy is used for the
boot source detection logic, but not yet for the barebox update code
which still uses the original information from IRAM.
Use the already existing rockchip_bootsource_get_active_slot() to
avoid data aborts when trying to detect the active boot slot.

Link: https://lore.barebox.org/20230509045055.601772-1-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>