ARM: i.MX6: TQMa6x: make tqma6x_env_init reusable
The SPI flash and eMMC are hard-wired on the TQMa6x SoM, so this
function is not specific for the mba6x baseboard but can be used on
other boards with a TQMa6x SoM too.

tqma6x_enet_init is specific to the mba6x board, so rename it to reflect
that fact.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Link: https://lore.barebox.org/20230511085803.335993-4-rhi@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>