video: bochs: fix IOMEM usage
For the PCI version of the driver, we end up calling IOMEM() twice on
both the fb_map and mmio pointers. This happens because we first use
pci_iomap() on PCI resources in the bochs_pci_probe() and then
explicitly call IOMEM() in the bochs_hw_probe(). However, judging from
the parameters of the latter function having __iomem attributes, we
should not be remapping them. So, remove those IOMEM calls and instead
do explicit remapping in bochs_isa_detect(), which was missing it.

Also fix bochs_hw_probe() declaration missing one __iomem.

Signed-off-by: Denis Orlov <denorl2009@gmail.com>
Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230511124743.2878952-1-denorl2009@gmail.com
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>