ARM: define stack base consistently
The different arm_mem_* functions have the pattern that they take
the region above it and substract the size of the current region. follow
the pattern for getting the stack base as well. While at it move
arm_mem_stack_top() lower in the file so that we have all functions
following said pattern below each other.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>