ARM: move arm_mem_scratch_get() lower for consistency
The different arm_mem_* functions all follow the same pattern of
taking the base address of the upper region minus the size of the
current region and with the exception of arm_mem_scratch_get() they
are all below each other. arm_mem_scratch_get() doesn't fit into
this row, so move it lower.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>