pbl: have DEBUG_PBL depend only on PBL_CONSOLE
PBL_CONSOLE enables use of printf and friends in the prebootloader. The
underlying putc will be what was passed to pbl_set_putc or alternatively
the DEBUG_LL PUTC_LL implementation.

In either case, PBL_CONSOLE is required to get the PBL output useful to
debugging, so let's drop the DEBUG_LL depends, but keep the help text to
lessen user confusion.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230522052222.1037931-5-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>