.clang-format: import kernel coding style rules from Linux v6.3
Make onboarding a bit easier by providing a clang-format ruleset
that formats code according to the kernel coding style that barebox
adheres to. The ruleset imported here from Linux v6.3 is not 100%
complete, but it's an ok starting point.

Only difference is adjustment of ForEachMacros to those defined by
barebox. Knowledge of more macros can be added later.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230522134210.2561620-3-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>