test: self: mmu: fix invalid use of __is_defined
__is_defined is only usable for undefined macros and defined macros with
a value. __is_defined on a defined macro without a value are will yield
zero, which is unfortunate as ARCH_HAS_REMAP is defined without a value
for the architectures that define it. The header has an arch_can_remap()
helper that's already used to guard remapping code in the memtest
command, so let's make use of it instead. While at it, fix the skipped
test count to correctly reflect the number of tests skipped.

This should fix failure of the test on x86, where remapping is not
supported as we run under EFI and don't touch page tables.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230523094208.3453910-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>