ARM: dts: rk3588s: Set CPU clock to 816MHz
Originally the Kernel rk3588s.dtsi assigned the CPU clock rates in the
scmi_clk node. We had to remove the assigned-clock-rates properties
because our SCMI clock driver didn't work properly. Now it does, so we
can bring these properties back. In the meantime the assigned-clock-rates
properties were moved to the cpu nodes in the upstream dtsi file in
87810bda8a847 ("arm64: dts: rockchip: Fix SCMI assigned clocks on rk3588s")
We do not have a driver on the CPU nodes (and neither does the kernel),
so the properties do not help us there. Instead, add the
assigned-clock-rates properties to our local rk3588s.dtsi file.

With this the CPU clocks are configured to 816MHz where the SoC
previously came up with only 24MHz.

Note that the CPU could run with 2.4GHz, but I am not sure yet if all
voltages are configured correctly for that frequency, so for now run
with the frequency that has been the default in the dtsi files.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>