imx-usb-loader: fix compilation if prepare make target was not called before
As pointed out in [1], compilation failed if the prepare make
target was not called before compiling the imx-usb-loader, which caused
the generated file not to exist when building for the sandbox target and
broke the build.

After a discussion on the original thread, a different approach has been pursued
by simply adding utsrelease.h as a dependency to scripts.

[1] https://lore.barebox.org/barebox/20230525181320.2277320-1-ahmad@a3f.at/

Fixes: c1b50061f4b3 ("imx-usb-loader: add commandline option for displaying version number")
Reported-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Johannes Zink <j.zink@pengutronix.de>
Tested-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230526165947.1079052-2-j.zink@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>