Revert "dma: use dma/cpu conversions correctly in dma_map/unmap_single"
This reverts commit d13d870986eeecc58d8652268557e4a159b9d4c4.

While the patch itself is correct, it at least breaks USB on the
Raspberry Pi 3b.

With this patch dma_sync_single_for_device() is passed the DMA address.
This is correct as even the prototype suggests that it should get a
dma_addr_t. Unfortunately this is not what the function implements and
also not what the users expect. Most if not all users simply pass a CPU
pointer casted to unsigned long. dma_sync_single_for_device() on ARM
then takes the DMA address as a CPU pointer and does cache maintenance
on it.

Before we can merge this patch again dma_sync_single_for_device() must
get a struct device * argument and (on ARM) the cpu_to_dma() conversion
must be reverted before doing cache maintenance.

Link: https://lore.barebox.org/20230531102337.843892-1-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>