test/Containerfile: reduce duplication in cross toolchain handling
ENV directive in OCI will result in the container having the environment
variable defined. Let's use that instead of duplicating it across
workflows.

Also the many wget's to get the kernel toolchains are not that readable,
so define a function and call it to make it easier to see what is what.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230605063623.1254764-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>