test/Containerfile: install only one RISC-V toolchain
riscv32-linux- and riscv64-linux- seem to only differ in defaults and
both are capable of compiling rv32 and rv64 code and indeed only one
toolchain was actually used.

It's probably less surprising to use rv64 toolchain for both rv32 and
v64, so let's keep only that and drop the non-unused rv32 toolchain.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230605063623.1254764-3-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>