partitions: efi: Continue partition enumeration on invalid pte
The efi partition enumeration stops, if a partition entry is
invalid or non-existent.
Later partitions which would be valid are therefore not enumerated.
This can be the case, eg. if a device with the following
partitions is enumerated: p1, p2, p4, p5, p6.
Skipping the invalid entry instead of exiting the enumeration fixes
this problem.

Signed-off-by: Christopher Ebner <ch.ebner@commend.com>
Link: https://lore.barebox.org/20230605101105.485283-1-ch.ebner@commend.com
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>