MIPS: add qemu-malta64el_defconfig
This defconfig will compile barebox for Malta board and MIPS64 5KEc/5KEf
CPUs, emulatable in QEMU. The analogous 32-bit defconfig uses big endian
CPU as the default one. So, for the sake of variety, and to improve the
coverage of options, set the CPU as little endian for this one.

Signed-off-by: Denis Orlov <denorl2009@gmail.com>
Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230607051155.3175-2-denorl2009@gmail.com
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>