ARM: replace CONFIG_CPU_V8 with CONFIG_CPU_64 in Makefile
We use CONFIG_CPU_V8 and CONFIG_CPU_64 interchangeably in barebox,
because we don't yet support any SoCs with ARMv9 and because we always
target the A64 instruction set on ARMv8 processors.

In the Makefile where we have decisions dependent on whether we are
compiling for 64-bit or not, we should rather use CONFIG_CPU_64 though,
because the same options would likely be applicable to ARMv9 as well.

Given that the follow-up commit will introduce a number of
CONFIG_CPU_32, let's replace the CONFIG_CPU_V8 symbols with
CONFIG_CPU_64 for symmetry. No functional change.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230607153152.2681009-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>