ARM: don't assume 32-bit when no boards are selected
barebox build errors are very confusing if no board is selected.
This should have been fixed with commit 14b296d2a7e6 ("arm: error
out if __LINUX_ARM_ARCH__ is undefined"), but unfortunately that's
only true for ARM32. On ARM64, the error message in question
is not printed, because build aborts even earlier, because Kbuild
assumes it should build for 32-bit ARM and thus passes the ARM64
compiler options that it can't understand:

  aarch64-oe-linux-gcc: error: unrecognized argument in option '-mabi=apcs-gnu'
  aarch64-oe-linux-gcc: note: valid arguments to '-mabi=' are: ilp32 lp64
  aarch64-oe-linux-gcc: error: unrecognized command-line option '-msoft-float'
  aarch64-oe-linux-gcc: error: unrecognized command-line option '-mno-unaligned-access'

Let's fix that for ARM64 builds by not assuming !CONFIG_CPU_64 to be 32-bit,
but instead explicitly check that CONFIG_CPU_32 is set before doing
32-bit specific changes.

This ensures we now fail during compilation on both ARM32 and ARM64 if
no boards were selected. We can't fail earlier via $(error ...) as this
would impact use of targets like menuconfig.

Reported-by: Lior Weintraub <liorw@pliops.com>
Reported-by: Ulrich Ölmann <u.oelmann@pengutronix.de>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230607153152.2681009-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>