ARM: dts: state-example: fix use with Phytec phyFLEX i.MX6
The state-example.dtsi was added 5 years ago and while the binding is
still applicable, the imx6q-phytec-pbab01.dts device tree that it
extends when USE_STATE_EXAMPLE is defined has changed in a way that
breaks use of the state nodes.

Do the necessary to restore the examples to working order:

  - Reference the NOR partition inside of the fixed-partition container
    it's inside now

  - Resize the NAND, so the last 1M of the barebox 16M partition can
    be used for state. Also use the fixed-partition container that
    has been added in the meantime.

  - Use phandles or label-relative paths, so we get build errors
    if the path changes again in future

There's still a WARNING: state_nor: Failed to initialize desired amount
of buckets, only 2 of 3 succeeded, but that can be fixed separately.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230608065805.3095109-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>